+++ CHANGELOG.md
@@ -1,3 +1,3 @@
 ## v0.1.0
 
-- Adds `coolify-resource-reconciler` package
+- Adds `coolify-deploy` package